split Racr and Test
